requirements/local.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
Django==4.1.2
djangorestframework==3.14.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.4
pytz==2022.4
sqlparse==0.4.3
tomli==2.0.1
